Move upload handler from PoliceCaseFiles component to CaseFiles component
